#20 - refactor: Type 패키지를 Product 패키지와 병합
`Type`은 `Product` 와 강한 연관이 있고
다대다 관계를 가지고 있어 분리하기보다
같은 패키지로 관리하는 것이 좋다고 판단하여 패키지를 병합함.
