+++ src/test/java/io/jeidiiy/sirenordersystem/product/controller/ProductControllerTest.java
@@ -7,12 +7,11 @@ import io.jeidiiy.sirenordersystem.jwt.service.JwtAuthenticationEntryPoint;
 import io.jeidiiy.sirenordersystem.jwt.service.JwtLogoutSuccessHandler;
 import io.jeidiiy.sirenordersystem.jwt.service.JwtService;
 import io.jeidiiy.sirenordersystem.product.domain.Product;
-import io.jeidiiy.sirenordersystem.product.domain.beverage.Beverage;
 import io.jeidiiy.sirenordersystem.product.domain.dto.ProductResponseDto;
 import io.jeidiiy.sirenordersystem.product.service.ProductService;
-import io.jeidiiy.sirenordersystem.type.domain.Category;
-import io.jeidiiy.sirenordersystem.type.domain.Type;
-import io.jeidiiy.sirenordersystem.type.domain.TypeProduct;
+import io.jeidiiy.sirenordersystem.product.domain.Category;
+import io.jeidiiy.sirenordersystem.product.domain.Type;
+import io.jeidiiy.sirenordersystem.product.domain.TypeProduct;
 import io.jeidiiy.sirenordersystem.user.service.UserService;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
