+++ src/test/java/io/jeidiiy/sirenordersystem/product/service/ProductServiceTest.java
@@ -7,9 +7,9 @@ import static org.mockito.BDDMockito.*;
 import io.jeidiiy.sirenordersystem.product.domain.Product;
 import io.jeidiiy.sirenordersystem.product.domain.dto.ProductResponseDto;
 import io.jeidiiy.sirenordersystem.product.repository.ProductJpaRepository;
-import io.jeidiiy.sirenordersystem.type.domain.Category;
-import io.jeidiiy.sirenordersystem.type.domain.Type;
-import io.jeidiiy.sirenordersystem.type.domain.TypeProduct;
+import io.jeidiiy.sirenordersystem.product.domain.Category;
+import io.jeidiiy.sirenordersystem.product.domain.Type;
+import io.jeidiiy.sirenordersystem.product.domain.TypeProduct;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
